+++ sdk/core/azure-core-amqp/src/models/private/value_impl.hpp
@@ -3,16 +3,17 @@
 
 #pragma once
 
-#include <azure/core/internal/unique_handle.hpp>
+#include "../../amqp/private/unique_handle.hpp"
 
 #include <azure_uamqp_c/amqpvalue.h>
 
-namespace Azure { namespace Core { namespace _internal {
+namespace Azure { namespace Core { namespace Amqp { namespace _detail {
   template <> struct UniqueHandleHelper<std::remove_pointer<AMQP_VALUE>::type>
   {
     // @cond INTERNAL
     static void FreeAmqpValue(AMQP_VALUE value);
-    using type = BasicUniqueHandle<std::remove_pointer<AMQP_VALUE>::type, FreeAmqpValue>;
+    using type
+        = Core::_internal::BasicUniqueHandle<std::remove_pointer<AMQP_VALUE>::type, FreeAmqpValue>;
     // @endcond
   };
 
@@ -20,18 +21,16 @@ namespace Azure { namespace Core { namespace _internal {
   {
     // @cond INTERNAL
     static void FreeAmqpDecoder(AMQPVALUE_DECODER_HANDLE value);
-    using type
-        = BasicUniqueHandle<std::remove_pointer<AMQPVALUE_DECODER_HANDLE>::type, FreeAmqpDecoder>;
+    using type = Core::_internal::
+        BasicUniqueHandle<std::remove_pointer<AMQPVALUE_DECODER_HANDLE>::type, FreeAmqpDecoder>;
     // @endcond
   };
-
-}}} // namespace Azure::Core::_internal
+}}}} // namespace Azure::Core::Amqp::_detail
 
 namespace Azure { namespace Core { namespace Amqp { namespace Models { namespace _detail {
-  using UniqueAmqpValueHandle
-      = Azure::Core::_internal::UniqueHandle<std::remove_pointer<AMQP_VALUE>::type>;
+  using UniqueAmqpValueHandle = Amqp::_detail::UniqueHandle<std::remove_pointer<AMQP_VALUE>::type>;
   using UniqueAmqpDecoderHandle
-      = Azure::Core::_internal::UniqueHandle<std::remove_pointer<AMQPVALUE_DECODER_HANDLE>::type>;
+      = Amqp::_detail::UniqueHandle<std::remove_pointer<AMQPVALUE_DECODER_HANDLE>::type>;
 
   class AmqpValueFactory final {
   public: