+++ sdk/storage/azure-storage-common/src/storage_per_retry_policy.cpp
@@ -3,6 +3,9 @@
 
 #include "azure/storage/common/internal/storage_per_retry_policy.hpp"
 
+#include "azure/storage/common/internal/constants.hpp"
+#include "azure/storage/common/internal/reliable_stream.hpp"
+
 #include <azure/core/datetime.hpp>
 #include <azure/core/platform.hpp>
 
@@ -16,9 +19,6 @@ namespace Azure { namespace Storage { namespace _internal {
       Core::Http::Policies::NextHttpPolicy nextPolicy,
       Core::Context const& context) const
   {
-    const char* HttpHeaderDate = "Date";
-    const char* HttpHeaderXMsDate = "x-ms-date";
-
     const auto& headers = request.GetHeaders();
     if (headers.find(HttpHeaderDate) == headers.end())
     {
@@ -29,11 +29,10 @@ namespace Azure { namespace Storage { namespace _internal {
               .ToString(Azure::DateTime::DateFormat::Rfc1123));
     }
 
-    const char* HttpHeaderTimeout = "timeout";
     auto cancelTimepoint = context.GetDeadline();
     if (cancelTimepoint == Azure::DateTime::max())
     {
-      request.GetUrl().RemoveQueryParameter(HttpHeaderTimeout);
+      request.GetUrl().RemoveQueryParameter(HttpQueryTimeout);
     }
     else
     {
@@ -43,8 +42,18 @@ namespace Azure { namespace Storage { namespace _internal {
                 .count()
           : -1;
       request.GetUrl().AppendQueryParameter(
-          HttpHeaderTimeout, std::to_string(std::max(numSeconds, int64_t(1))));
+          HttpQueryTimeout, std::to_string(std::max(numSeconds, int64_t(1))));
     }
+
+    std::string client_request_id;
+    if (context.TryGetValue(ReliableStreamClientRequestIdKey, client_request_id))
+    {
+      if (!client_request_id.empty())
+      {
+        request.SetHeader(HttpHeaderClientRequestId, client_request_id);
+      }
+    }
+
     return nextPolicy.Send(request, context);
   }